chore: reorganize unit tests for apps-v5 package
