+++ packages/apps-v5/test/unit/commands/apps/destroy.unit.test.js
