+++ packages/apps-v5/test/unit/commands/apps/errors.unit.test.js
