+++ packages/apps-v5/test/unit/commands/apps/favorites/add.unit.test.js
