+++ packages/apps-v5/test/unit/commands/apps/favorites/index.unit.test.js
