+++ packages/apps-v5/test/unit/commands/apps/favorites/remove.unit.test.js
