+++ packages/apps-v5/test/unit/commands/apps/info.unit.test.js
