+++ packages/apps-v5/test/unit/commands/apps/rename.unit.test.js
