+++ packages/apps-v5/test/unit/commands/features/enable.unit.test.js
