+++ packages/apps-v5/test/unit/commands/features/info.unit.test.js
