+++ packages/apps-v5/test/unit/commands/keys/clear.unit.test.js
