+++ packages/apps-v5/test/unit/commands/releases/info.unit.test.js
